+++ NEWS.md
@@ -11,7 +11,7 @@ Known issues: <https://github.com/PredictiveEcology/SpaDES.core/issues>
 * `params` and `.globals` were previously not expected to change during `Cache`d events. Thus returned cached values were always the same as input as `params` and `.globals`. They are now assessed and returned as part of the `Cache`, as expected.
 * updates to handle upstream changes in `Require` and `reproducible`, including renaming `cacheRepo` to `cachePath` in some inherited functions.
 * updates to sample modules to use `SpaDES.tools::neutralLandscapeMap` instead of `NLMR` package directly
-* migration complete to use `terra` and `sf` instead of `raster`, `sp`, `rgeos`, and `rgdal` as defaults. Attemps have been made to maintain backwards compatibility in all cases.
+* migration complete to use `terra` and `sf` instead of `raster`, `sp`, `rgeos`, and `rgdal` as defaults. Attempts have been made to maintain backwards compatibility in all cases.
 * `moduleMetadata` now handles multiple module paths
 * updates to `memoryUse`
 * new option: setting `options("spades.allowInitDuringSimInit" = TRUE)`, a user will have `init` events of *one ore more* modules run during the `simInit` call, but only if they have no upstream dependencies, i.e., their `expectsInputs` cannot be supplied by another module's `createsOutputs`. `simInit` will determine which modules have no upstream dependencies and *only* these will be selected for running *only* their `init` events. This can be useful e.g., if there is a module that `createsOutputs` for a `studyArea`.